Door lock switch added
Added a door lock switch, which stops the door from opening when active. The doors were given a nav mesh obstacle attribute to be an obstacle, but they have weird behavior, meaning that the feature is a bit wonky ATM. It may be removed later on, as it is not an essential feature.